+++ HISTORY.md
@@ -6,6 +6,8 @@
 
 ### mlpack 3.2.0
 ###### 2019-09-25
+  * Fix some potential infinity errors in Naive Bayes Classifier (#2022).
+
   * Fix occasionally-failing RADICAL test (#1924).
 
   * Fix gcc 9 OpenMP compilation issue (#1970).